fix: Address review — deduplicate page objects, remove redundant waitFor
- Move TopologyPage methods into existing topology-page.ts, remove
  duplicate class from add-page.ts
- Merge SoftwareCatalogPage into existing catalog-page.ts, rename to
  CatalogPage, fix import from fixtures
- Remove redundant waitFor() before robustClick in selectResourceType
  and selectRuntimeIcon — pass timeout to robustClick instead
- Remove duplicate saveButton/clickSave from DeployImagePage
- Update imports in 6 spec files
